+++ src/init.c
@@ -2459,13 +2459,33 @@ SEXP savvy_PlRExpr_str_slice__impl(SEXP self__, SEXP c_arg__offset, SEXP c_arg__
     return handle_result(res);
 }
 
-SEXP savvy_PlRExpr_str_split__impl(SEXP self__, SEXP c_arg__by, SEXP c_arg__inclusive) {
-    SEXP res = savvy_PlRExpr_str_split__ffi(self__, c_arg__by, c_arg__inclusive);
+SEXP savvy_PlRExpr_str_split__impl(SEXP self__, SEXP c_arg__by) {
+    SEXP res = savvy_PlRExpr_str_split__ffi(self__, c_arg__by);
     return handle_result(res);
 }
 
-SEXP savvy_PlRExpr_str_split_exact__impl(SEXP self__, SEXP c_arg__by, SEXP c_arg__n, SEXP c_arg__inclusive) {
-    SEXP res = savvy_PlRExpr_str_split_exact__ffi(self__, c_arg__by, c_arg__n, c_arg__inclusive);
+SEXP savvy_PlRExpr_str_split_exact__impl(SEXP self__, SEXP c_arg__by, SEXP c_arg__n) {
+    SEXP res = savvy_PlRExpr_str_split_exact__ffi(self__, c_arg__by, c_arg__n);
+    return handle_result(res);
+}
+
+SEXP savvy_PlRExpr_str_split_exact_inclusive__impl(SEXP self__, SEXP c_arg__by, SEXP c_arg__n) {
+    SEXP res = savvy_PlRExpr_str_split_exact_inclusive__ffi(self__, c_arg__by, c_arg__n);
+    return handle_result(res);
+}
+
+SEXP savvy_PlRExpr_str_split_inclusive__impl(SEXP self__, SEXP c_arg__by) {
+    SEXP res = savvy_PlRExpr_str_split_inclusive__ffi(self__, c_arg__by);
+    return handle_result(res);
+}
+
+SEXP savvy_PlRExpr_str_split_regex__impl(SEXP self__, SEXP c_arg__by, SEXP c_arg__strict) {
+    SEXP res = savvy_PlRExpr_str_split_regex__ffi(self__, c_arg__by, c_arg__strict);
+    return handle_result(res);
+}
+
+SEXP savvy_PlRExpr_str_split_regex_inclusive__impl(SEXP self__, SEXP c_arg__by, SEXP c_arg__strict) {
+    SEXP res = savvy_PlRExpr_str_split_regex_inclusive__ffi(self__, c_arg__by, c_arg__strict);
     return handle_result(res);
 }
 
@@ -3940,8 +3960,12 @@ static const R_CallMethodDef CallEntries[] = {
     {"savvy_PlRExpr_str_replace_many__impl", (DL_FUNC) &savvy_PlRExpr_str_replace_many__impl, 5},
     {"savvy_PlRExpr_str_reverse__impl", (DL_FUNC) &savvy_PlRExpr_str_reverse__impl, 1},
     {"savvy_PlRExpr_str_slice__impl", (DL_FUNC) &savvy_PlRExpr_str_slice__impl, 3},
-    {"savvy_PlRExpr_str_split__impl", (DL_FUNC) &savvy_PlRExpr_str_split__impl, 3},
-    {"savvy_PlRExpr_str_split_exact__impl", (DL_FUNC) &savvy_PlRExpr_str_split_exact__impl, 4},
+    {"savvy_PlRExpr_str_split__impl", (DL_FUNC) &savvy_PlRExpr_str_split__impl, 2},
+    {"savvy_PlRExpr_str_split_exact__impl", (DL_FUNC) &savvy_PlRExpr_str_split_exact__impl, 3},
+    {"savvy_PlRExpr_str_split_exact_inclusive__impl", (DL_FUNC) &savvy_PlRExpr_str_split_exact_inclusive__impl, 3},
+    {"savvy_PlRExpr_str_split_inclusive__impl", (DL_FUNC) &savvy_PlRExpr_str_split_inclusive__impl, 2},
+    {"savvy_PlRExpr_str_split_regex__impl", (DL_FUNC) &savvy_PlRExpr_str_split_regex__impl, 3},
+    {"savvy_PlRExpr_str_split_regex_inclusive__impl", (DL_FUNC) &savvy_PlRExpr_str_split_regex_inclusive__impl, 3},
     {"savvy_PlRExpr_str_splitn__impl", (DL_FUNC) &savvy_PlRExpr_str_splitn__impl, 3},
     {"savvy_PlRExpr_str_starts_with__impl", (DL_FUNC) &savvy_PlRExpr_str_starts_with__impl, 2},
     {"savvy_PlRExpr_str_strip_chars__impl", (DL_FUNC) &savvy_PlRExpr_str_strip_chars__impl, 2},