+++ NEWS.md
@@ -19,6 +19,7 @@ This is an update that corresponds to Python Polars 1.38.1.
   (`^...$`) and wildcards (`*`) in column names are expanded
   ([pola-rs/polars#26437](https://github.com/pola-rs/polars/pull/26437)).
 - New `<expr>$bin$get()` to extract a specific byte from a binary value (#1731).
+- `<expr>$str$split()` has two new arguments `literal` and `strict` (#1730).
 
 ### Bug fixes